feat(balance): mostrar el saldo del usuario
