+++ src/views/Home.vue
@@ -1,16 +1,25 @@
 <template>
   <div class="home row justify-content-center">
-    <BillingTable></BillingTable>
+    <div class="col-12">
+      <div class="float-right">
+        <Balance></Balance>
+      </div>
+    </div>
+    <div class="col-12">
+      <BillingTable></BillingTable>
+    </div>
   </div>
 </template>
 
 <script>
 import BillingTable from "../components/BillingTable";
+import Balance from "../components/Balance";
 
 export default {
   name: 'Home',
   components: {
-    BillingTable
+    BillingTable,
+    Balance
   }
 }
 </script>
